requirements.txt
# Core FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23

# ML & Embeddings
sentence-transformers==2.3.1
transformers==4.36.0
torch==2.1.0
faiss-cpu==1.7.4
rank-bm25==0.2.2
numpy==1.24.3
scikit-learn==1.3.2
scipy==1.11.4

# NLP & Tokenization
tiktoken==0.5.1

# HTTP client for Ollama
httpx==0.25.2

# Config & Validation
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
